requirements.txt
django-cors-headers==2.1.0
django-extensions==2.0.7
django-import-export==1.2.0
django-storages==1.6.6
django-utils==0.0.2
djangorestframework==3.8.2
mysqlclient==1.4.4
Pillow==4.1.1
oauth_valid==0.2.3
drf-yasg==1.15.0
ktis-parser==0.3.1
Django==2.1
